feat(rust): create standalone commands for kafka services
