+++ implementations/rust/ockam/ockam_command/src/service/start.rs
@@ -1,28 +1,15 @@
-use std::str::FromStr;
-
 use crate::node::NodeOpts;
 use crate::util::{api, node_rpc, RpcBuilder};
 use crate::Result;
 use crate::{fmt_warn, CommandGlobalOpts};
 use anyhow::anyhow;
 use clap::{Args, Subcommand};
-use colorful::Colorful;
+
 use minicbor::Encode;
 use ockam::{Context, TcpTransport};
-use ockam_api::nodes::models::services::{
-    StartKafkaConsumerRequest, StartKafkaProducerRequest, StartServiceRequest,
-};
-use ockam_api::port_range::PortRange;
-use ockam_api::DefaultAddress;
-use ockam_core::api::{Request, RequestBuilder, Status};
-use ockam_core::compat::net::{Ipv4Addr, SocketAddr};
-use ockam_multiaddr::MultiAddr;
 
-const KAFKA_DEFAULT_PROJECT_ROUTE: &str = "/project/default";
-const KAFKA_DEFAULT_CONSUMER_SERVER: &str = "127.0.0.1:4000";
-const KAFKA_DEFAULT_CONSUMER_PORT_RANGE: &str = "4001-4100";
-const KAFKA_DEFAULT_PRODUCER_SERVER: &str = "127.0.0.1:5000";
-const KAFKA_DEFAULT_PRODUCER_PORT_RANGE: &str = "5001-5100";
+use ockam_api::DefaultAddress;
+use ockam_core::api::{RequestBuilder, Status};
 
 /// Start a specified service
 #[derive(Clone, Debug, Args)]
@@ -69,38 +56,6 @@ pub enum StartSubCommand {
         #[arg(long)]
         project: String,
     },
-    KafkaConsumer {
-        /// The local address of the service
-        #[arg(long, default_value_t = kafka_consumer_default_addr())]
-        addr: String,
-        /// The address where to bind and where the client will connect to alongside its port, <address>:<port>.
-        /// In case just a port is specified, the default loopback address (127.0.0.1) will be used
-        #[arg(long, default_value_t = kafka_default_consumer_server(), value_parser = parse_bootstrap_server)]
-        bootstrap_server: SocketAddr,
-        /// Local port range dynamically allocated to kafka brokers, must not overlap with the
-        /// bootstrap port
-        #[arg(long, default_value_t = kafka_default_consumer_port_range())]
-        brokers_port_range: PortRange,
-        /// The route to the project in ockam orchestrator, expected something like /project/<name>
-        #[arg(long, default_value_t = kafka_default_project_route())]
-        project_route: MultiAddr,
-    },
-    KafkaProducer {
-        /// The local address of the service
-        #[arg(long, default_value_t = kafka_producer_default_addr())]
-        addr: String,
-        /// The address where to bind and where the client will connect to alongside its port, <address>:<port>.
-        /// In case just a port is specified, the default loopback address (127.0.0.1) will be used
-        #[arg(long, default_value_t = kafka_default_producer_server(), value_parser = parse_bootstrap_server)]
-        bootstrap_server: SocketAddr,
-        /// Local port range dynamically allocated to kafka brokers, must not overlap with the
-        /// bootstrap port
-        #[arg(long, default_value_t = kafka_default_producer_port_range())]
-        brokers_port_range: PortRange,
-        /// The route to the project in ockam orchestrator, expected something like /project/<name>
-        #[arg(long, default_value_t = kafka_default_project_route())]
-        project_route: MultiAddr,
-    },
 }
 
 fn vault_default_addr() -> String {
@@ -125,38 +80,6 @@ fn credentials_default_addr() -> String {
 
 fn authenticator_default_addr() -> String {
     DefaultAddress::DIRECT_AUTHENTICATOR.to_string()
-}
-
-fn kafka_consumer_default_addr() -> String {
-    DefaultAddress::KAFKA_CONSUMER.to_string()
-}
-
-fn kafka_producer_default_addr() -> String {
-    DefaultAddress::KAFKA_PRODUCER.to_string()
-}
-
-fn kafka_default_project_route() -> MultiAddr {
-    MultiAddr::from_str(KAFKA_DEFAULT_PROJECT_ROUTE).expect("Failed to parse default project route")
-}
-
-fn kafka_default_consumer_server() -> SocketAddr {
-    SocketAddr::from_str(KAFKA_DEFAULT_CONSUMER_SERVER)
-        .expect("Failed to parse default consumer server")
-}
-
-fn kafka_default_consumer_port_range() -> PortRange {
-    PortRange::from_str(KAFKA_DEFAULT_CONSUMER_PORT_RANGE)
-        .expect("Failed to parse default consumer port range")
-}
-
-fn kafka_default_producer_server() -> SocketAddr {
-    SocketAddr::from_str(KAFKA_DEFAULT_PRODUCER_SERVER)
-        .expect("Failed to parse default producer server")
-}
-
-fn kafka_default_producer_port_range() -> PortRange {
-    PortRange::from_str(KAFKA_DEFAULT_PRODUCER_PORT_RANGE)
-        .expect("Failed to parse default producer port range")
 }
 
 impl StartCommand {
@@ -214,72 +137,13 @@ async fn run_impl(
         StartSubCommand::Authenticator { addr, project, .. } => {
             start_authenticator_service(ctx, &opts, node_name, &addr, &project, Some(&tcp)).await?
         }
-        StartSubCommand::KafkaConsumer {
-            addr,
-            bootstrap_server,
-            brokers_port_range,
-            project_route,
-        } => {
-            let payload =
-                StartKafkaConsumerRequest::new(bootstrap_server, brokers_port_range, project_route);
-            let payload = StartServiceRequest::new(payload, &addr);
-            let req = Request::post("/node/services/kafka_consumer").body(payload);
-
-            opts.terminal.write_line(&fmt_warn!(
-                "Starting KafkaConsumer service at {}",
-                &bootstrap_server.to_string()
-            ))?;
-            opts.terminal.write_line(&fmt_warn!(
-                "Brokers port range set to {}",
-                &brokers_port_range.to_string()
-            ))?;
-            start_service_impl(
-                ctx,
-                &opts,
-                node_name,
-                &addr,
-                "KafkaConsumer",
-                req,
-                Some(&tcp),
-            )
-            .await?
-        }
-        StartSubCommand::KafkaProducer {
-            addr,
-            bootstrap_server,
-            brokers_port_range,
-            project_route,
-        } => {
-            let payload =
-                StartKafkaProducerRequest::new(bootstrap_server, brokers_port_range, project_route);
-            let payload = StartServiceRequest::new(payload, &addr);
-            let req = Request::post("/node/services/kafka_producer").body(payload);
-            opts.terminal.write_line(&fmt_warn!(
-                "Starting KafkaProducer service at {}",
-                &bootstrap_server.to_string()
-            ))?;
-            opts.terminal.write_line(&fmt_warn!(
-                "Brokers port range set to {}",
-                &brokers_port_range.to_string()
-            ))?;
-            start_service_impl(
-                ctx,
-                &opts,
-                node_name,
-                &addr,
-                "KafkaProducer",
-                req,
-                Some(&tcp),
-            )
-            .await?
-        }
     }
 
     Ok(())
 }
 
 /// Helper function.
-async fn start_service_impl<T>(
+pub(crate) async fn start_service_impl<T>(
     ctx: &Context,
     opts: &CommandGlobalOpts,
     node_name: &str,
@@ -304,34 +168,6 @@ where
             eprintln!("{}", rpc.parse_err_msg(res, dec));
             Err(anyhow!("Failed to start {serv_name} service").into())
         }
-    }
-}
-
-/// Helper routine for parsing bootstrap server ip and port from user input
-/// It can parse a string containing either an `ip:port` pair or just a `port`
-/// into a valid SocketAddr instance.
-fn parse_bootstrap_server(bootstrap_server: &str) -> Result<SocketAddr> {
-    let addr: Vec<&str> = bootstrap_server.split(':').collect();
-    match addr.len() {
-        // Only the port is available
-        1 => {
-            let port: u16 = addr[0]
-                .parse()
-                .map_err(|_| anyhow!("Invalid port number"))?;
-            let ip: Ipv4Addr = [127, 0, 0, 1].into();
-            Ok(SocketAddr::new(ip.into(), port))
-        }
-        // Both the ip and port are available
-        2 => {
-            let port: u16 = addr[1]
-                .parse()
-                .map_err(|_| anyhow!("Invalid port number"))?;
-            let ip = addr[0]
-                .parse::<Ipv4Addr>()
-                .map_err(|_| anyhow!("Invalid IP address"))?;
-            Ok(SocketAddr::new(ip.into(), port))
-        }
-        _ => Err(anyhow!("Failed to parse bootstrap server").into()),
     }
 }
 
@@ -383,48 +219,4 @@ pub async fn start_authenticator_service(
 ) -> Result<()> {
     let req = api::start_authenticator_service(serv_addr, project);
     start_service_impl(ctx, opts, node_name, serv_addr, "Authenticator", req, tcp).await
-}
-
-#[cfg(test)]
-mod tests {
-    use ockam_core::compat::net::{IpAddr, Ipv4Addr, SocketAddr};
-
-    use crate::service::start::parse_bootstrap_server;
-
-    #[test]
-    fn test_parse_bootstrap_server() {
-        // Test case 1: only a port is provided
-        let input = "9000";
-        let result = parse_bootstrap_server(input);
-        assert!(result.is_ok());
-        if let Ok(bootstrap_server) = result {
-            assert_eq!(
-                SocketAddr::new(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)), 9000),
-                bootstrap_server
-            );
-        }
-
-        // Test case 2: Any 4 octet combination (IPv4) followed by ":" like in "192.168.0.1:9999"
-        let input = "192.168.0.1:9999";
-        let result = parse_bootstrap_server(input);
-        assert!(result.is_ok());
-        if let Ok(bootstrap_server) = result {
-            assert_eq!(
-                SocketAddr::new(IpAddr::V4(Ipv4Addr::new(192, 168, 0, 1)), 9999),
-                bootstrap_server
-            );
-        }
-
-        // Test case 3: Any other format will throw an error
-        let invalid_input = "invalid";
-        assert!(parse_bootstrap_server(invalid_input).is_err());
-
-        let invalid_input = "192.168.0.1:invalid";
-        assert!(parse_bootstrap_server(invalid_input).is_err());
-
-        let invalid_input = "192.168.0.1:9999:extra";
-        assert!(parse_bootstrap_server(invalid_input).is_err());
-        let invalid_input = "192,166,0.1:9999";
-        assert!(parse_bootstrap_server(invalid_input).is_err());
-    }
 }
